requirements.txt
pyautogui
pillow